requirements.txt
Flask
requests
pytesseract
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
PyPDF2
pdf2image
Pillow
python-docx
langchain
langchain-core
langchain-community
langchain-openai
streamlit
werkzeug
gunicorn
gspread
google-auth
